Fix: test with jest
