+++ api/static/.~lock.test.xlsx#
@@ -1,0 +1,1 @@
+,GWTN141-10/Franc,GWTN141-10,06.11.2023 13:12,file:///C:/Users/Franc/AppData/Roaming/LibreOffice/4;
