+++ project/assets/test.csv
@@ -1,0 +1,5 @@
+ASIN,TITLE,COSTUNIT,LISTPRICE,QUANTITY,PURCHASEDDATE,SUPPLIER,CONDITION,MSKU,SALESTAX,DISCOUNT,EXPIRATION DATE,NOTES,TAXCODE,MINPRICE,MAXPRICE
+B09XZXRFPX,,100,1,2,11/1/23,amazon.com,,11/1/23-$100-Standard-White,4,,,,A_GEN_TAX,200,1
+B0BY1HQBSX,,100,20,3,11/1/23,amazon.com,,11/1/23-$100-Standard-White,4,,,,A_GEN_TAX,5,20
+B0C1P7XV7V,,100,185,3,11/1/23,amazon.com,,11/1/23-$100-Universal-Black,4,,,,A_GEN_TAX,129.6,185
+B07XF2TGGK,,100,101,2,11/1/23,amazon.com,,11/1/23-$100--White,4,,,,A_GEN_TAX,130.5,101
