+++ project/package.json
@@ -37,5 +37,10 @@
     "@types/jest": "^29.5.7",
     "babel-jest": "^29.7.0",
     "ts-jest": "^29.1.1"
+  },
+  "static": {
+    "includes": [
+      "assets"
+    ]
   }
 }
